backup/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiohttp==3.8.6
opencv-python==4.8.1.78
numpy==1.24.3
schedule==1.2.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
aiosqlite==0.19.0
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
schedule
